Fixed profile page picture, company and displayed username name to be pulled from db
