+++ fwb/app/(groups)/groups/page.tsx
@@ -2,7 +2,8 @@ import { DiscountData } from "@/app/types/types";
 import Navbar from "@/components/ui/privategroups/groups_navbar";
 import GroupDetailsSection from "@/components/ui/privategroups/GroupDetailsSection";
 import Tabs from "@/components/ui/privategroups/Tabs";
-import { auth } from "@clerk/nextjs";import { Container, Box } from "@mui/material";
+import { auth } from "@clerk/nextjs";
+import { Container, Box } from "@mui/material";
 import { relative } from "path";
 import SearchBar from "@/components/ui/privategroups/SearchBar";
